refactor: refactor bfd static route tests
