+++ tests/bfd/test_bfd_static_route.py
@@ -1,15 +1,17 @@
-import pytest
-from bfd_base import BfdBase
 import logging
 import time
 
+import pytest
+
+from tests.bfd.bfd_base import BfdBase
 from tests.bfd.bfd_helpers import verify_static_route, select_src_dst_dut_with_asic, control_interface_state, \
-    check_bgp_status
-from tests.common.utilities import wait_until
+    check_bgp_status, modify_all_bfd_sessions, find_bfd_peers_with_given_state, add_bfd, verify_bfd_state, delete_bfd, \
+    extract_backend_portchannels
 from tests.common.config_reload import config_reload
+from tests.common.fixtures.tacacs import tacacs_creds, setup_tacacs  # noqa F401
 from tests.common.platform.processes_utils import wait_critical_processes
 from tests.common.reboot import reboot
-from tests.common.fixtures.tacacs import tacacs_creds, setup_tacacs    # noqa F401
+from tests.common.utilities import wait_until
 
 pytestmark = [pytest.mark.topology("t2")]
 
@@ -20,7 +22,7 @@ class TestBfdStaticRoute(BfdBase):
     TOTAL_ITERATIONS = 100
 
     @pytest.fixture(autouse=True, scope="class")
-    def modify_bfd_sessions(self, duthosts, bfd_base_instance):
+    def modify_bfd_sessions(self, duthosts):
         """
         1. Gather all front end nodes
         2. Modify BFD state to required state & issue config reload.
@@ -35,7 +37,7 @@ class TestBfdStaticRoute(BfdBase):
         try:
             duts = duthosts.frontend_nodes
             for dut in duts:
-                bfd_base_instance.modify_all_bfd_sessions(dut, "false")
+                modify_all_bfd_sessions(dut, "false")
             for dut in duts:
                 # config reload
                 config_reload(dut)
@@ -50,7 +52,7 @@ class TestBfdStaticRoute(BfdBase):
                         600,
                         10,
                         0,
-                        lambda: bfd_base_instance.find_bfd_peers_with_given_state(
+                        lambda: find_bfd_peers_with_given_state(
                             dut, asic, "No BFD sessions found"
                         ),
                     )
@@ -60,7 +62,7 @@ class TestBfdStaticRoute(BfdBase):
         finally:
             duts = duthosts.frontend_nodes
             for dut in duts:
-                bfd_base_instance.modify_all_bfd_sessions(dut, "true")
+                modify_all_bfd_sessions(dut, "true")
             for dut in duts:
                 config_reload(dut)
                 wait_critical_processes(dut)
@@ -74,29 +76,26 @@ class TestBfdStaticRoute(BfdBase):
                         600,
                         10,
                         0,
-                        lambda: bfd_base_instance.find_bfd_peers_with_given_state(
+                        lambda: find_bfd_peers_with_given_state(
                             dut, asic, "Up"
                         ),
                     )
 
-    def test_bfd_with_lc_reboot_ipv4(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_with_lc_reboot(
         self,
         localhost,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv4"
-
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
         logger.info(
             "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
@@ -111,22 +110,22 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -134,7 +133,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -148,19 +147,14 @@ class TestBfdStaticRoute(BfdBase):
         # Waiting for all processes on Source dut
         wait_critical_processes(src_dut)
 
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -168,14 +162,14 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Savings the configs
         src_dut.shell("sudo config save -y")
@@ -186,19 +180,14 @@ class TestBfdStaticRoute(BfdBase):
         # Waiting for all processes on Source dut
         wait_critical_processes(src_dut)
 
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -206,152 +195,20 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
 
-    def test_bfd_with_lc_reboot_ipv6(
-        self,
-        localhost,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv6"
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-
-        # Perform a cold reboot on source dut
-        reboot(src_dut, localhost)
-
-        # Waiting for all processes on Source dut
-        wait_critical_processes(src_dut)
-
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-
-        # Config reload of Source dut
-        reboot(src_dut, localhost)
-
-        # Waiting for all processes on Source dut
-        wait_critical_processes(src_dut)
-
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-
-    def test_bfd_static_route_deletion_ipv4(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_static_route_deletion(
         self,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
@@ -365,8 +222,6 @@ class TestBfdStaticRoute(BfdBase):
             4. Delete BFD on Destination dut.
             5. Verify that on Destination dut BFD gets cleaned up and static route will be added back.
         """
-        version = "ipv4"
-
         logger.info(
             "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
         )
@@ -380,21 +235,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -402,20 +257,20 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD & Static route verifications")
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
             ),
         )
@@ -423,7 +278,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
@@ -433,7 +288,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -442,19 +296,18 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
 
         logger.info("BFD deletion on destination dut")
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         logger.info("BFD & Static route verifications")
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -462,7 +315,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
@@ -472,7 +325,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -481,158 +333,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
 
         logger.info("BFD deletion did not influence static routes and test completed successfully")
 
-    def test_bfd_static_route_deletion_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_flap(
         self,
         duthost,
         request,
         duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv6"
-
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD & Static route verifications")
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-
-        logger.info("BFD deletion on destination dut")
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        logger.info("BFD & Static route verifications")
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-
-        logger.info("BFD deletion did not influence static routes and test completed successfully")
-
-    def test_bfd_flap_ipv4(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
@@ -648,8 +363,6 @@ class TestBfdStaticRoute(BfdBase):
             6. Verify that on destination dut BFD is up and static route is added back.
             7. Repeat above steps 100 times.
         """
-        version = "ipv4"
-
         logger.info(
             "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
         )
@@ -663,21 +376,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -685,7 +398,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -695,7 +408,7 @@ class TestBfdStaticRoute(BfdBase):
             logger.info("Iteration {}".format(i))
 
             logger.info("BFD deletion on source dut")
-            bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+            delete_bfd(src_asic.asic_index, src_prefix, src_dut)
 
             logger.info("Waiting for 5s post BFD shutdown")
             time.sleep(5)
@@ -705,7 +418,7 @@ class TestBfdStaticRoute(BfdBase):
                 180,
                 10,
                 0,
-                lambda: bfd_base_instance.verify_bfd_state(
+                lambda: verify_bfd_state(
                     dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
                 ),
             )
@@ -713,7 +426,7 @@ class TestBfdStaticRoute(BfdBase):
                 180,
                 10,
                 0,
-                lambda: bfd_base_instance.verify_bfd_state(
+                lambda: verify_bfd_state(
                     src_dut,
                     src_dut_nexthops.values(),
                     src_asic,
@@ -726,7 +439,6 @@ class TestBfdStaticRoute(BfdBase):
                 dst_prefix,
                 dst_dut,
                 "Route Removal",
-                bfd_base_instance,
                 version,
             )
             verify_static_route(
@@ -735,19 +447,18 @@ class TestBfdStaticRoute(BfdBase):
                 src_prefix,
                 src_dut,
                 "Route Addition",
-                bfd_base_instance,
                 version,
             )
 
             logger.info("BFD addition on source dut")
-            bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+            add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
             logger.info("BFD & Static route verifications")
             assert wait_until(
                 180,
                 10,
                 0,
-                lambda: bfd_base_instance.verify_bfd_state(
+                lambda: verify_bfd_state(
                     dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
                 ),
             )
@@ -755,7 +466,7 @@ class TestBfdStaticRoute(BfdBase):
                 180,
                 10,
                 0,
-                lambda: bfd_base_instance.verify_bfd_state(
+                lambda: verify_bfd_state(
                     src_dut, src_dut_nexthops.values(), src_asic, "Up"
                 ),
             )
@@ -765,7 +476,6 @@ class TestBfdStaticRoute(BfdBase):
                 dst_prefix,
                 dst_dut,
                 "Route Addition",
-                bfd_base_instance,
                 version,
             )
             verify_static_route(
@@ -774,7 +484,6 @@ class TestBfdStaticRoute(BfdBase):
                 src_prefix,
                 src_dut,
                 "Route Addition",
-                bfd_base_instance,
                 version,
             )
 
@@ -793,176 +502,8 @@ class TestBfdStaticRoute(BfdBase):
 
         logger.info("test_bfd_flap completed")
 
-    def test_bfd_flap_ipv6(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-
-        To flap the BFD session ( Up <--> Down <---> Up) between linecards for 100 times.
-            Test Steps:
-            1. Delete BFD on Source dut
-            2. Verify that on Source dut BFD gets cleaned up and static route exists.
-            3. Verify that on Destination dut BFD goes down and static route will be removed.
-            4. Add BFD on Source dut.
-            5. Verify that on Source dut BFD is up
-            6. Verify that on destination dut BFD is up and static route is added back.
-            7. Repeat above steps 100 times.
-        """
-        version = "ipv6"
-
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        successful_iterations = 0  # Counter for successful iterations
-        for i in range(self.TOTAL_ITERATIONS):
-            logger.info("Iteration {}".format(i))
-
-            logger.info("BFD deletion on source dut")
-            bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-            logger.info("Waiting for 5s post BFD shutdown")
-            time.sleep(5)
-
-            logger.info("BFD & Static route verifications")
-            assert wait_until(
-                180,
-                10,
-                0,
-                lambda: bfd_base_instance.verify_bfd_state(
-                    dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
-                ),
-            )
-            assert wait_until(
-                180,
-                10,
-                0,
-                lambda: bfd_base_instance.verify_bfd_state(
-                    src_dut,
-                    src_dut_nexthops.values(),
-                    src_asic,
-                    "No BFD sessions found",
-                ),
-            )
-            verify_static_route(
-                request,
-                dst_asic,
-                dst_prefix,
-                dst_dut,
-                "Route Removal",
-                bfd_base_instance,
-                version,
-            )
-            verify_static_route(
-                request,
-                src_asic,
-                src_prefix,
-                src_dut,
-                "Route Addition",
-                bfd_base_instance,
-                version,
-            )
-
-            logger.info("BFD addition on source dut")
-            bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-            logger.info("BFD & Static route verifications")
-            assert wait_until(
-                180,
-                10,
-                0,
-                lambda: bfd_base_instance.verify_bfd_state(
-                    dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-                ),
-            )
-            assert wait_until(
-                180,
-                10,
-                0,
-                lambda: bfd_base_instance.verify_bfd_state(
-                    src_dut, src_dut_nexthops.values(), src_asic, "Up"
-                ),
-            )
-            verify_static_route(
-                request,
-                dst_asic,
-                dst_prefix,
-                dst_dut,
-                "Route Addition",
-                bfd_base_instance,
-                version,
-            )
-            verify_static_route(
-                request,
-                src_asic,
-                src_prefix,
-                src_dut,
-                "Route Addition",
-                bfd_base_instance,
-                version,
-            )
-
-            # Check if both iterations were successful and increment the counter
-            successful_iterations += 1
-
-        # Determine the success rate
-        logger.info("successful_iterations: %d", successful_iterations)
-        success_rate = (successful_iterations / self.TOTAL_ITERATIONS) * 100
-
-        logger.info("Current success rate: %.2f%%", success_rate)
-        # Check if the success rate is above the threshold (e.g., 98%)
-        assert (
-            success_rate >= 98
-        ), "BFD flap verification success rate is below 98% ({}%)".format(success_rate)
-
-        logger.info("test_bfd_flap completed")
-
-    def test_bfd_with_rp_reboot_ipv4(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_with_rp_reboot(
         self,
         localhost,
         duthost,
@@ -970,17 +511,14 @@ class TestBfdStaticRoute(BfdBase):
         duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         enum_supervisor_dut_hostname,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv4"
-
         rp = duthosts[enum_supervisor_dut_hostname]
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -997,22 +535,22 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1020,7 +558,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1036,19 +574,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1056,14 +589,14 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Savings the configs
         src_dut.shell("sudo config save -y")
@@ -1076,19 +609,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -1096,167 +624,25 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
 
-    def test_bfd_with_rp_reboot_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_remote_link_flap(
         self,
-        localhost,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        enum_supervisor_dut_hostname,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv6"
-
-        rp = duthosts[enum_supervisor_dut_hostname]
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Perform a cold reboot on source dut
-        reboot(rp, localhost)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Config reload of Source dut
-        reboot(rp, localhost)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-
-    def test_bfd_remote_link_flap_ipv4(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv4"
-
         request.config.interface_shutdown = True
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -1273,21 +659,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1295,7 +681,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1317,7 +703,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Down"
             ),
         )
@@ -1330,7 +716,6 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
 
@@ -1345,7 +730,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1353,7 +738,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1366,7 +751,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -1375,27 +759,23 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
 
-    def test_bfd_remote_link_flap_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_lc_asic_shutdown(
         self,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv6"
-
         request.config.interface_shutdown = True
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -1412,21 +792,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1434,146 +814,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Extract portchannel interfaces on dst
-        list_of_portchannels_on_dst = src_dut_nexthops.keys()
-        request.config.portchannels_on_dut = "dst"
-        request.config.selected_portchannels = list_of_portchannels_on_dst
-
-        # Shutdown PortChannels on destination dut
-        for interface in list_of_portchannels_on_dst:
-            action = "shutdown"
-            control_interface_state(
-                dst_dut, dst_asic, interface, action
-            )
-
-        # Verification of BFD session state on src dut
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Down"
-            ),
-        )
-
-        # Verify that corresponding static route has been removed on both duts
-        logger.info("BFD & Static route verifications")
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-
-        for interface in list_of_portchannels_on_dst:
-            action = "startup"
-            control_interface_state(
-                dst_dut, dst_asic, interface, action
-            )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Verify that corresponding static route has been added on both duts
-        logger.info("BFD & Static route verifications")
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-
-    def test_bfd_lc_asic_shutdown_ipv4(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv4"
-
-        request.config.interface_shutdown = True
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1595,7 +836,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
             ),
         )
@@ -1603,7 +844,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Down"
             ),
         )
@@ -1616,7 +857,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -1625,7 +865,6 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
 
@@ -1640,7 +879,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1648,7 +887,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1661,7 +900,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -1670,27 +908,23 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
 
-    def test_bfd_lc_asic_shutdown_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_portchannel_member_flap(
         self,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv6"
-
         request.config.interface_shutdown = True
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -1707,21 +941,21 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1729,163 +963,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Extract portchannel interfaces on src
-        list_of_portchannels_on_src = dst_dut_nexthops.keys()
-        request.config.portchannels_on_dut = "src"
-        request.config.selected_portchannels = list_of_portchannels_on_src
-
-        # Shutdown PortChannels
-        for interface in list_of_portchannels_on_src:
-            action = "shutdown"
-            control_interface_state(
-                src_dut, src_asic, interface, action
-            )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Down"
-            ),
-        )
-
-        # Verify that corresponding static route has been removed on both duts
-        logger.info("BFD & Static route verifications")
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-
-        for interface in list_of_portchannels_on_src:
-            action = "startup"
-            control_interface_state(
-                src_dut, src_asic, interface, action
-            )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Verify that corresponding static route has been added on both duts
-        logger.info("BFD & Static route verifications")
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-
-    def test_bfd_portchannel_member_flap_ipv4(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv4"
-
-        request.config.interface_shutdown = True
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1899,7 +977,7 @@ class TestBfdStaticRoute(BfdBase):
         for portchannel_interface in list_of_portchannels_on_src:
             action = "shutdown"
             list_of_portchannel_members_on_src = (
-                bfd_base_instance.extract_backend_portchannels(src_dut)[
+                extract_backend_portchannels(src_dut)[
                     portchannel_interface
                 ]["members"]
             )
@@ -1916,7 +994,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
             ),
         )
@@ -1924,7 +1002,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Down"
             ),
         )
@@ -1937,7 +1015,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -1946,7 +1023,6 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Removal",
-            bfd_base_instance,
             version,
         )
 
@@ -1954,7 +1030,7 @@ class TestBfdStaticRoute(BfdBase):
         for portchannel_interface in list_of_portchannels_on_src:
             action = "startup"
             list_of_portchannel_members_on_src = (
-                bfd_base_instance.extract_backend_portchannels(src_dut)[
+                extract_backend_portchannels(src_dut)[
                     portchannel_interface
                 ]["members"]
             )
@@ -1968,7 +1044,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -1976,7 +1052,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -1989,7 +1065,6 @@ class TestBfdStaticRoute(BfdBase):
             dst_prefix,
             dst_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
         verify_static_route(
@@ -1998,29 +1073,23 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             src_dut,
             "Route Addition",
-            bfd_base_instance,
             version,
         )
 
-    def test_bfd_portchannel_member_flap_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_config_reload(
         self,
         duthost,
         request,
-        duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv6"
-
-        request.config.interface_shutdown = True
-
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
         logger.info(
             "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
@@ -2035,192 +1104,22 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Extract portchannel interfaces on src
-        list_of_portchannels_on_src = dst_dut_nexthops.keys()
-        request.config.portchannels_on_dut = "src"
-        request.config.selected_portchannels = list_of_portchannels_on_src
-
-        # Shutdown PortChannel members
-        for portchannel_interface in list_of_portchannels_on_src:
-            action = "shutdown"
-            list_of_portchannel_members_on_src = (
-                bfd_base_instance.extract_backend_portchannels(src_dut)[
-                    portchannel_interface
-                ]["members"]
-            )
-            request.config.selected_portchannel_members = (
-                list_of_portchannel_members_on_src
-            )
-            for each_member in list_of_portchannel_members_on_src:
-                control_interface_state(
-                    src_dut, src_asic, each_member, action
-                )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
-            ),
-        )
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Down"
-            ),
-        )
-
-        # Verify that corresponding static route has been removed on both duts
-        logger.info("BFD & Static route verifications")
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Removal",
-            bfd_base_instance,
-            version,
-        )
-
-        # Bring up of PortChannel members
-        for portchannel_interface in list_of_portchannels_on_src:
-            action = "startup"
-            list_of_portchannel_members_on_src = (
-                bfd_base_instance.extract_backend_portchannels(src_dut)[
-                    portchannel_interface
-                ]["members"]
-            )
-            for each_member in list_of_portchannel_members_on_src:
-                control_interface_state(
-                    src_dut, src_asic, each_member, action
-                )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Verify that corresponding static route has been added on both duts
-        logger.info("Static route verifications")
-        verify_static_route(
-            request,
-            dst_asic,
-            dst_prefix,
-            dst_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-        verify_static_route(
-            request,
-            src_asic,
-            src_prefix,
-            src_dut,
-            "Route Addition",
-            bfd_base_instance,
-            version,
-        )
-
-    def test_bfd_config_reload_ipv4(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv4"
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2228,7 +1127,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -2242,19 +1141,14 @@ class TestBfdStaticRoute(BfdBase):
         # Waiting for all processes on Source dut
         wait_critical_processes(src_dut)
 
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2262,14 +1156,14 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Savings the configs
         src_dut.shell("sudo config save -y")
@@ -2280,19 +1174,14 @@ class TestBfdStaticRoute(BfdBase):
         # Waiting for all processes on Source dut
         wait_critical_processes(src_dut)
 
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -2300,143 +1189,13 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
 
-    def test_bfd_config_reload_ipv6(
-        self,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv6"
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-
-        # Config reload of Source dut
-        config_reload(src_dut)
-
-        # Waiting for all processes on Source dut
-        wait_critical_processes(src_dut)
-
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-
-        # Config reload of Source dut
-        config_reload(src_dut)
-
-        # Waiting for all processes on Source dut
-        wait_critical_processes(src_dut)
-
-        assert wait_until(
-            300,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-
-    def test_bfd_with_rp_config_reload_ipv4(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_with_rp_config_reload(
         self,
         localhost,
         duthost,
@@ -2444,17 +1203,14 @@ class TestBfdStaticRoute(BfdBase):
         duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         enum_supervisor_dut_hostname,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv4"
-
         rp = duthosts[enum_supervisor_dut_hostname]
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -2471,22 +1227,22 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2494,7 +1250,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -2510,19 +1266,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2530,14 +1281,14 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Savings the configs
         src_dut.shell("sudo config save -y")
@@ -2550,18 +1301,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
+
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -2569,12 +1316,13 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
 
-    def test_bfd_with_rp_config_reload_ipv6(
+    @pytest.mark.parametrize("version", ["ipv4", "ipv6"])
+    def test_bfd_with_bad_fc_asic(
         self,
         localhost,
         duthost,
@@ -2582,17 +1330,14 @@ class TestBfdStaticRoute(BfdBase):
         duthosts,
         tbinfo,
         get_src_dst_asic_and_duts,
-        bfd_base_instance,
         enum_supervisor_dut_hostname,
         bfd_cleanup_db,
+        version,
     ):
         """
         Author:  Harsha Golla
         Email : harsgoll@cisco.com
         """
-
-        version = "ipv6"
-
         rp = duthosts[enum_supervisor_dut_hostname]
 
         # Selecting source, destination dut & prefix & BFD status verification for all nexthops
@@ -2609,22 +1354,22 @@ class TestBfdStaticRoute(BfdBase):
             src_prefix,
             dst_prefix,
         ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
+            request, get_src_dst_asic_and_duts, version
         )
 
         # Creation of BFD
         logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
+        add_bfd(src_asic.asic_index, src_prefix, src_dut)
 
         logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Verification of BFD session state.
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2632,146 +1377,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Perform a cold reboot on source dut
-        config_reload(rp)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source & destination dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Config reload of Source dut
-        config_reload(rp)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-
-    def test_bfd_with_bad_fc_asic_ipv4(
-        self,
-        localhost,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        enum_supervisor_dut_hostname,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv4"
-
-        rp = duthosts[enum_supervisor_dut_hostname]
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
@@ -2787,15 +1393,15 @@ class TestBfdStaticRoute(BfdBase):
         asic_ids = [int(element.split("swss")[1]) for element in docker_output]
 
         # Shut down corresponding asic on supervisor to simulate bad asic
-        for id in asic_ids:
-            rp.shell("systemctl stop swss@{}".format(id))
+        for asic_id in asic_ids:
+            rp.shell("systemctl stop swss@{}".format(asic_id))
 
         # Verify that BFD sessions are down
         assert wait_until(
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
             ),
         )
@@ -2803,7 +1409,7 @@ class TestBfdStaticRoute(BfdBase):
             180,
             10,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Down"
             ),
         )
@@ -2815,19 +1421,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
             ),
         )
@@ -2835,14 +1436,14 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "Up"
             ),
         )
 
         logger.info("BFD deletion on source dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
+        delete_bfd(src_asic.asic_index, src_prefix, src_dut)
+        delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
 
         # Savings the configs
         src_dut.shell("sudo config save -y")
@@ -2855,19 +1456,14 @@ class TestBfdStaticRoute(BfdBase):
         wait_critical_processes(src_dut)
         wait_critical_processes(dst_dut)
 
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
+        check_bgp_status(request)
 
         # Verification of BFD session state.
         assert wait_until(
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
             ),
         )
@@ -2875,174 +1471,7 @@ class TestBfdStaticRoute(BfdBase):
             300,
             20,
             0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
-            ),
-        )
-
-    def test_bfd_with_bad_fc_asic_ipv6(
-        self,
-        localhost,
-        duthost,
-        request,
-        duthosts,
-        tbinfo,
-        get_src_dst_asic_and_duts,
-        bfd_base_instance,
-        enum_supervisor_dut_hostname,
-        bfd_cleanup_db,
-    ):
-        """
-        Author:  Harsha Golla
-        Email : harsgoll@cisco.com
-        """
-
-        version = "ipv6"
-
-        rp = duthosts[enum_supervisor_dut_hostname]
-
-        # Selecting source, destination dut & prefix & BFD status verification for all nexthops
-        logger.info(
-            "Selecting Source dut, destination dut, source asic, destination asic, source prefix, destination prefix"
-        )
-        (
-            src_asic,
-            dst_asic,
-            src_dut,
-            dst_dut,
-            src_dut_nexthops,
-            dst_dut_nexthops,
-            src_prefix,
-            dst_prefix,
-        ) = select_src_dst_dut_with_asic(
-            request, get_src_dst_asic_and_duts, bfd_base_instance, version
-        )
-
-        # Creation of BFD
-        logger.info("BFD addition on source dut")
-        bfd_base_instance.add_bfd(src_asic.asic_index, src_prefix, src_dut)
-
-        logger.info("BFD addition on destination dut")
-        bfd_base_instance.add_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Verification of BFD session state.
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Extract asic ids
-        docker_output = rp.shell("docker ps | grep swss | awk '{print $NF}'")[
-            "stdout"
-        ].split("\n")
-        asic_ids = [int(element.split("swss")[1]) for element in docker_output]
-
-        # Shut down corresponding asic on supervisor to simulate bad asic
-        for id in asic_ids:
-            rp.shell("systemctl stop swss@{}".format(id))
-
-        # Verify that BFD sessions are down
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Down"
-            ),
-        )
-        assert wait_until(
-            180,
-            10,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Down"
-            ),
-        )
-
-        # Config reload RP to bring up the swss containers
-        config_reload(rp)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "Up"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                src_dut, src_dut_nexthops.values(), src_asic, "Up"
-            ),
-        )
-
-        logger.info("BFD deletion on source dut")
-        bfd_base_instance.delete_bfd(src_asic.asic_index, src_prefix, src_dut)
-        bfd_base_instance.delete_bfd(dst_asic.asic_index, dst_prefix, dst_dut)
-
-        # Savings the configs
-        src_dut.shell("sudo config save -y")
-        dst_dut.shell("sudo config save -y")
-
-        # Config reload RP
-        config_reload(rp)
-
-        # Waiting for all processes on Source & destination dut
-        wait_critical_processes(src_dut)
-        wait_critical_processes(dst_dut)
-
-        assert wait_until(
-            600,
-            10,
-            0,
-            lambda: check_bgp_status(request),
-        )
-
-        # Verification of BFD session state.
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
-                dst_dut, dst_dut_nexthops.values(), dst_asic, "No BFD sessions found"
-            ),
-        )
-        assert wait_until(
-            300,
-            20,
-            0,
-            lambda: bfd_base_instance.verify_bfd_state(
+            lambda: verify_bfd_state(
                 src_dut, src_dut_nexthops.values(), src_asic, "No BFD sessions found"
             ),
         )
